feat: add calendar header component
